requirements.txt
bcrypt # 4.0.1
beanie # 1.18.0
fastapi # 0.95.0
httpx # 0.24.0
mongomock_motor # 0.0.19
motor #3.1.1
passlib # 1.7.4
pydantic
pydantic[email,dotenv] #1.10.7
PyJWT
pymongo # 4.3.3
pytest # 7.3.1
pytest-mock # 3.10.0
python-dotenv
python-jose
uvicorn # 0.21.1